refactor: 현재 패키지들 product 패키지 하위로 이동
